+++ marketdata-insight-service/marketdata-insight-application/src/main/java/shop/shportfolio/marketdata/insight/application/dto/candle/response/CandleDayResponseDto.java
@@ -4,6 +4,8 @@ package shop.shportfolio.marketdata.insight.application.dto.candle.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class CandleDayResponseDto {
@@ -11,15 +13,15 @@ public class CandleDayResponseDto {
     private final String market;
     private final String candleDateTimeUtc;
     private final String candleDateTimeKst;
-    private final Double openingPrice;
-    private final Double highPrice;
-    private final Double lowPrice;
-    private final Double tradePrice;
-    private final Double candleAccTradePrice;
-    private final Double candleAccTradeVolume;
-    private final Double prevClosingPrice;
-    private final Double changePrice;
-    private final Double changeRate;
+    private final BigDecimal openingPrice;
+    private final BigDecimal highPrice;
+    private final BigDecimal lowPrice;
+    private final BigDecimal tradePrice;
+    private final BigDecimal candleAccTradePrice;
+    private final BigDecimal candleAccTradeVolume;
+    private final BigDecimal prevClosingPrice;
+    private final BigDecimal changePrice;
+    private final BigDecimal changeRate;
 
     @Override
     public String toString() {
