+++ marketdata-insight-service/marketdata-insight-application/src/main/java/shop/shportfolio/marketdata/insight/application/dto/candle/response/CandleMinuteResponseDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class CandleMinuteResponseDto {
@@ -11,13 +13,13 @@ public class CandleMinuteResponseDto {
     private final String marketId;
     private final String candleDateTimeUTC;
     private final String candleDateTimeKST;
-    private final Double openingPrice;
-    private final Double highPrice;
-    private final Double lowPrice;
-    private final Double tradePrice;
+    private final BigDecimal openingPrice;
+    private final BigDecimal highPrice;
+    private final BigDecimal lowPrice;
+    private final BigDecimal tradePrice;
     private final Long timestamp;
-    private final Double candleAccTradePrice;
-    private final Double candleAccTradeVolume;
+    private final BigDecimal candleAccTradePrice;
+    private final BigDecimal candleAccTradeVolume;
     private final Integer unit;
 
     @Override
